Finish migration to new db access scheme.
